Fixed for FF
As the use of Manifest V2 is only useful in Firefox I updated this folder so that it works well on Firefox browser. 

This means that I had to add the fonts locally 
and that I had to change the padding/margin of certain blocks in the popup, because otherwise it will not get displayed normally.