C: Fix popup scroll bar in Firefox
Because the popup had a fixed `width` but some elements inside the popup had to large margins it caused in Firefox to display a useless scrollbar.  
Instead using `min/max-width` and `display: block;` (and not a very large margin to create a colorful background) may solve this problem.  
Although the Microsoft Edge browser handled this inconsistency better, this new way of coding is better. Thus I applied the changes for the Firefox extension also to the Edge extension.